requirements.txt
pyodbc
pandas
azure-identity
python-dotenv
sqlalchemy
azure-storage-blob
pyarrow
pillow
openpyxl